- Fix 'utlis' -> 'utils' typo in folder and all import statements - Move formatLifespan to dateHandlers.js - Remove  formatter.js utility file - Update imports
